Make the test code as the tutorial code.

+++ theano/tests/test_tutorial.py
@@ -585,7 +585,7 @@ class T_examples(unittest.TestCase):
         from theano import shared
         # Force the dtype to int64 to work correctly on 32 bit computer.
         # Otherwise, it create by default a int32 on 32 bit computer.
-        state = shared(numpy.int64(0))
+        state = shared(0)
         inc = T.iscalar('inc')
         accumulator = function([inc], state, updates=[(state, state+inc)])
